# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_tmalldedede_agentbox_internal_provider.Provider:
    properties:
      agent:
        description: Agent compatibility
        type: string
      api_key_url:
        description: Where to get API key
        type: string
      base_url:
        description: API Configuration
        type: string
      category:
        allOf:
        - $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_provider.ProviderCategory'
        description: Category
      default_model:
        description: Model defaults
        type: string
      default_models:
        description: Available models
        items:
          type: string
        type: array
      description:
        description: Provider description
        type: string
      docs_url:
        description: Documentation URL
        type: string
      env_config:
        additionalProperties:
          type: string
        description: Environment variables template
        type: object
      icon:
        description: UI
        type: string
      icon_color:
        description: Icon color (hex)
        type: string
      id:
        description: Basic info
        type: string
      is_built_in:
        description: Flags
        type: boolean
      is_enabled:
        description: Is enabled
        type: boolean
      is_partner:
        description: Partner provider (featured)
        type: boolean
      name:
        description: e.g., "DeepSeek"
        type: string
      requires_ak:
        description: Requires API key
        type: boolean
      website_url:
        description: URLs
        type: string
    type: object
  github_com_tmalldedede_agentbox_internal_provider.ProviderCategory:
    enum:
    - official
    - cn_official
    - aggregator
    - third_party
    type: string
    x-enum-varnames:
    - CategoryOfficial
    - CategoryCNOfficial
    - CategoryAggregator
    - CategoryThirdParty
...
  github_com_tmalldedede_agentbox_internal_task.Input:
    properties:
      branch:
        type: string
      path:
        type: string
      text:
        type: string
      type:
        description: git / files / text
        type: string
      url:
        type: string
    type: object
  github_com_tmalldedede_agentbox_internal_task.OutputConfig:
    properties:
      files:
        description: 指定输出的文件
        items:
          type: string
        type: array
      format:
        description: 输出格式
        type: string
      type:
        description: files / text / json
        type: string
    type: object
  github_com_tmalldedede_agentbox_internal_task.OutputFile:
    properties:
      mime_type:
        type: string
      name:
        type: string
      path:
        type: string
      size:
        type: integer
      url:
        description: 下载地址
        type: string
    type: object
  github_com_tmalldedede_agentbox_internal_task.Result:
    properties:
      files:
        items:
          $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_task.OutputFile'
        type: array
      logs:
        type: string
      summary:
        type: string
      text:
        type: string
      usage:
        $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_task.Usage'
    type: object
  github_com_tmalldedede_agentbox_internal_task.Status:
    enum:
    - pending
    - queued
    - running
    - completed
    - failed
    - cancelled
    type: string
    x-enum-comments:
      StatusCancelled: 用户取消
      StatusCompleted: 执行成功
      StatusFailed: 执行失败
      StatusPending: 刚创建，等待入队
      StatusQueued: 已入队，等待调度
      StatusRunning: 容器启动，Agent 执行中
    x-enum-descriptions:
    - 刚创建，等待入队
    - 已入队，等待调度
    - 容器启动，Agent 执行中
    - 执行成功
    - 执行失败
    - 用户取消
    x-enum-varnames:
    - StatusPending
    - StatusQueued
    - StatusRunning
    - StatusCompleted
    - StatusFailed
    - StatusCancelled
  github_com_tmalldedede_agentbox_internal_task.Task:
    properties:
      agent_type:
        description: claude-code / codex
        type: string
      completed_at:
        type: string
      created_at:
        description: 时间戳
        type: string
      error_message:
        description: 失败原因
        type: string
      id:
        description: 基本信息
        type: string
      input:
        $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_task.Input'
      metadata:
        additionalProperties:
          type: string
        description: 元数据
        type: object
      output:
        allOf:
        - $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_task.OutputConfig'
        description: 输出配置
      profile_id:
        type: string
      profile_name:
        description: 冗余，方便展示
        type: string
      prompt:
        description: 任务内容
        type: string
      queued_at:
        type: string
      result:
        allOf:
        - $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_task.Result'
        description: 执行结果
      session_id:
        description: 关联的 Session
        type: string
      started_at:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_task.Status'
        description: 运行时状态
      timeout:
        description: 执行配置
        type: integer
      webhook_url:
        type: string
    type: object
  github_com_tmalldedede_agentbox_internal_task.Usage:
    properties:
      duration_seconds:
        type: integer
      input_tokens:
        type: integer
      output_tokens:
        type: integer
      total_tokens:
        type: integer
    type: object
...
  internal_api.CreateProviderRequest:
    description: Request body for creating a new provider
    properties:
      agent:
        type: string
      api_key_url:
        type: string
      base_url:
        type: string
      category:
        type: string
      default_model:
        type: string
      default_models:
        items:
          type: string
        type: array
      description:
        type: string
      docs_url:
        type: string
      env_config:
        additionalProperties:
          type: string
        type: object
      icon:
        type: string
      icon_color:
        type: string
      id:
        type: string
      name:
        type: string
      website_url:
        type: string
    required:
    - agent
    - id
    - name
    type: object
  internal_api.CreateTaskRequest:
    properties:
      input:
        $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_task.Input'
      metadata:
        additionalProperties:
          type: string
        type: object
      output:
        $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_task.OutputConfig'
      profile_id:
        type: string
      prompt:
        type: string
      session_id:
        description: 可选：使用已存在的 Session
        type: string
      timeout:
        type: integer
      webhook_url:
        type: string
    required:
    - profile_id
    - prompt
    type: object
...
  internal_api.ListTasksResponse:
    properties:
      tasks:
        items:
          $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_task.Task'
        type: array
      total:
        type: integer
    type: object
  internal_api.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      type:
        description: 错误类型
        type: string
    type: object
  internal_api.UpdateProviderRequest:
    description: Request body for updating an existing provider
    properties:
      base_url:
        type: string
      default_model:
        type: string
      default_models:
        items:
          type: string
        type: array
      description:
        type: string
      env_config:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
    type: object
...
host: localhost:18080
info:
  contact:
    email: support@agentbox.dev
    name: AgentBox Team
    url: https://github.com/user/agentbox
  description: |-
    AgentBox - Open-source AI Agent containerized runtime platform

    ## Overview
    AgentBox provides a unified API for managing AI coding agents (Claude Code, Codex, OpenCode) in containerized environments.

    ## Authentication
    Currently no authentication required. API keys are managed via Credentials API.

    ## API Structure
    - **Public API** (`/api/v1/*`): Core functionality for external integrations
    - **Admin API** (`/api/v1/admin/*`): Platform management operations
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: AgentBox API
  version: "1.0"
paths:
...
  /providers:
    get:
      consumes:
      - application/json
      description: Get a list of API providers with optional filtering by agent type
        or category
      parameters:
      - description: Filter by agent type
        enum:
        - claude-code
        - codex
        - opencode
        - all
        in: query
        name: agent
        type: string
      - description: Filter by category
        enum:
        - official
        - cn_official
        - aggregator
        - third_party
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_provider.Provider'
                  type: array
              type: object
      summary: List all providers
      tags:
      - Providers
    post:
      consumes:
      - application/json
      description: Create a new custom API provider configuration
      parameters:
      - description: Provider configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateProviderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_provider.Provider'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Create a provider
      tags:
      - Providers
  /providers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a custom provider (built-in providers cannot be deleted)
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Delete a provider
      tags:
      - Providers
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific provider
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_provider.Provider'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Get a provider
      tags:
      - Providers
    put:
      consumes:
      - application/json
      description: Update an existing provider configuration (custom providers only)
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.UpdateProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_provider.Provider'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Update a provider
      tags:
      - Providers
...
  /tasks:
    get:
      description: Get a list of all tasks with optional filtering and pagination
      parameters:
      - description: Filter by status
        enum:
        - pending
        - queued
        - running
        - completed
        - failed
        - cancelled
        in: query
        name: status
        type: string
      - description: Filter by Profile ID
        in: query
        name: profile_id
        type: string
      - default: 20
        description: Number of results to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_api.ListTasksResponse'
              type: object
      summary: List tasks
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Create a new async task for batch processing with a specified profile
      parameters:
      - description: Task configuration
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_task.Task'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Create a task
      tags:
      - Tasks
  /tasks/{id}:
    delete:
      description: Cancel a pending or running task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_task.Task'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Cancel a task
      tags:
      - Tasks
    get:
      description: Get detailed information about a specific task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_task.Task'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Get a task
      tags:
      - Tasks
  /tasks/{id}/logs:
    get:
      description: Get execution logs for a task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  properties:
                    logs:
                      type: string
                    status:
                      type: string
                    task_id:
                      type: string
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Get task logs
      tags:
      - Tasks
  /tasks/{id}/output:
    get:
      description: Get the result/output of a completed task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tmalldedede_agentbox_internal_task.Result'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Get task output
      tags:
      - Tasks
...
swagger: "2.0"
...